Added ability to edit employment on profile page
